New title screen background
The old background was anything but seamless, and it also used an old screenshot. This one uses modern shaders, the updated resource pack, and proper editing techniques (including making a cuboid screenshot and splitting it into 6 parts) to make a nicer one, because it's nice to have new things time to time.